Centralize GitHub token/repo helpers
Move duplicated get_github_token and get_repo_info logic into a shared
e2e/perf/github_common.py module and update perf scripts to import it.
